\name{disCRS}

\alias{disCRS}
 
\title{
Predictor variables
}

\description{
Get the coordinate reference system for the data of a region.  
}

\usage{
disCRS(region)
}
 
\arguments{
  \item{region}{character. One of "AWT", "CAN", "NSW", "NZ", "SA", "SWI" }
}


\value{
character vector
}

\examples{
disCRS("AWT")
disCRS("NSW")
}

\keyword{data}
\keyword{spatial}

