\name{disPredictors}

\alias{disPredictors}
 
\title{
Predictor variables
}

\description{
Get the names of the predictor variables for a region.  
}

\usage{
disPredictors(region)
}
 
\arguments{
  \item{region}{character. One of "AWT", "CAN", "NSW", "NZ", "SA", "SWI" }
}


\value{
character vector
}


\examples{
disPredictors("NSW")
}

\keyword{data}

