\name{predictors}

\alias{predictors}

 
\title{
predictor variables
}

\description{
Helper function to get the names of the predictor variables for a dataset.  
}

\usage{
predictors(region)
}
 
\arguments{
  \item{region}{Character. One of 'AWT', 'CAN', 'NSW', 'NZ', 'SA', 'SWI' }
}


\value{
character vector
}

\author{
Robert J. Hijmans
}

\examples{
predictors('NSW')
}

\keyword{data}

