\name{disdat-package}

\alias{disdat-package}
\alias{disdat}

\docType{package}

\title{
Species distribution modeling
}

\description{
This package consists of a collection of datasets that can be used to compare species distribution models. There are data for 6 regions in the world, for a total of xxx species. Each data set has presence-only training data to build models, and presence/absence data to evaluate models.

The data were compiled and used by a species distribution modelling working group sponsored by the National Center for Ecological Analysis and Synthesis (NCEAS), at UC Santa Barbara, USA. 
}

\details{
Longtude and latitude have the WGS84 datum.
}


\author{
Jane Elith, Catherine Graham, Robert J. Hijmans, Steven Phillips.
}

\references{
The main reference for the this data set is:

J. Elith, C.H. Graham, R.P. Anderson, M. Dudik, S. Ferrier, A. Guisan, R.J. Hijmans, F. Huettmann, J. Leathwick, A. Lehmann, J. Li, L.G. Lohmann, B. Loiselle, G. Manion, C. Moritz, M. Nakamura, Y. Nakazawa, J. McC. Overton, A.T. Peterson, S. Phillips, K. Richardson, R. Scachetti-Pereira, R. Schapire, J. Soberon, S. Williams, M. Wisz and N. Zimmerman, 2006. Novel methods improve prediction of species' distributions from occurrence data. Ecography 29: 129-151. \url{http://dx.doi.org/10.1111/j.2006.0906-7590.04596.x}


Other papers using these data include:

Catherine H. Graham, Jane Elith, Robert J. Hijmans, Antoine Guisan, A. Townsend Peterson, Bette A. Loiselle and the NCEAS Predicting Species Distributions Working Group, 2007. The influence of spatial errors in species occurrence data used in distribution models. Journal of Applied Ecology 45: 239-247

Antoine Guisan, Catherine H. Graham, Jane Elith, Falk Huettmann and the NCEAS Species Distribution Modelling Group, 2007. Sensitivity of predictive species distribution models to change in grain size. Diversity and Distributions 13: 332-340.

Steven J. Phillips, Miroslav Dudik, Jane Elith, Catherine H. Graham, Anthony Lehmann, John Leathwick, Simon Ferrier, 2009. Sample selection bias and presence-only distribution models: implications for background and pseudo-absence data. Ecological Applications 19: 181-197.
\url{http://dx.doi.org/10.1890/07-2153.1}

M.S. Wisz, R.J. Hijmans, J. Li, A.T. Peterson, C.H. Graham, A. Guisan, and the NCEAS Predicting Species Distributions Working Group, 2008. Effects of sample size on the performance of species distribution models. Diversity and Distributions 14: 763-773.
}

\keyword{ package }
\keyword{ spatial }

